+++ core/ross.h
@@ -208,10 +208,7 @@ typedef uint64_t tw_lpid;
 #  include "tw-memory.h"
 #endif
 
-#include "instrumentation/st-stats-buffer.h"
-#include "instrumentation/st-data-collection.h"
-#include "instrumentation/st-event-collection.h"
-#include "instrumentation/st-model-collection.h"
+#include "instrumentation/st-instrumentation.h"
 
 #include "tw-eventq.h"
