+++ core/tw-pe.c
@@ -71,33 +71,6 @@ tw_pe_init(tw_peid id, tw_peid gid)
 
 	g_tw_pe[id]->rng = tw_rand_init(31, 41);
 
-    // set up files and buffers for necessary instrumentation modes
-    if (g_st_stats_enabled)
-    {
-        char suffix[4];
-        sprintf(suffix, "gvt");
-        g_st_buffer_gvt = st_buffer_init(suffix, &g_st_gvt_fh);
-    }
-    if (g_st_real_time_samp)
-    {
-        g_st_real_time_samp = g_st_real_time_samp * g_tw_clock_rate / 1000;
-        g_st_real_samp_start_cycles = tw_clock_read();
-        char suffix[3];
-        sprintf(suffix, "rt");
-        g_st_buffer_rt = st_buffer_init(suffix, &g_st_rt_fh);
-    }
-    if (g_st_ev_trace)
-    {
-        char suffix[8];
-        sprintf(suffix, "evtrace");
-        g_st_buffer_evrb = st_buffer_init(suffix, &g_st_evrb_fh);
-    }
-    if (g_st_model_stats)
-    {
-        char suffix[6];
-        sprintf(suffix, "model");
-        g_st_buffer_model = st_buffer_init(suffix, &g_st_model_fh);
-    }
 }
 
 void
